@@ -177,7 +177,7 @@ func (opc *Controller) Run(ctx context.Context, workers int) {
 		return
 	}
 
-	for i := 0; i < workers; i++ {
+	for range workers {
 		opc.wg.Add(1)
 		go func() {
 			defer opc.wg.Done()
