@@ -592,7 +592,6 @@ func TestApplyStatefulSet(t *testing.T) {
 	}
 
 	for _, tc := range tt {
-		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
@@ -605,7 +604,7 @@ func TestApplyStatefulSet(t *testing.T) {
 			if tc.expectedErr != nil {
 				iterations = 1
 			}
-			for i := 0; i < iterations; i++ {
+			for i := range iterations {
 				t.Run("", func(t *testing.T) {
 					ctx, ctxCancel := context.WithTimeout(context.Background(), 30*time.Second)
 					defer ctxCancel()
@@ -1093,7 +1092,6 @@ func TestApplyDaemonSet(t *testing.T) {
 	}
 
 	for _, tc := range tt {
-		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
@@ -1106,7 +1104,7 @@ func TestApplyDaemonSet(t *testing.T) {
 			if tc.expectedErr != nil {
 				iterations = 1
 			}
-			for i := 0; i < iterations; i++ {
+			for i := range iterations {
 				t.Run("", func(t *testing.T) {
 					ctx, ctxCancel := context.WithTimeout(context.Background(), 30*time.Second)
 					defer ctxCancel()
