@@ -464,7 +464,6 @@ func TestApplyPodDisruptionBudget(t *testing.T) {
 	}
 
 	for _, tc := range tt {
-		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
@@ -477,7 +476,7 @@ func TestApplyPodDisruptionBudget(t *testing.T) {
 			if tc.expectedErr != nil {
 				iterations = 1
 			}
-			for i := 0; i < iterations; i++ {
+			for i := range iterations {
 				t.Run("", func(t *testing.T) {
 					ctx, ctxCancel := context.WithTimeout(context.Background(), 30*time.Second)
 					defer ctxCancel()
