@@ -411,7 +411,6 @@ func TestApplyClusterRole(t *testing.T) {
 	}
 
 	for _, tc := range tt {
-		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
@@ -424,7 +423,7 @@ func TestApplyClusterRole(t *testing.T) {
 			if tc.expectedErr != nil {
 				iterations = 1
 			}
-			for i := 0; i < iterations; i++ {
+			for i := range iterations {
 				t.Run("", func(t *testing.T) {
 					ctx, ctxCancel := context.WithTimeout(context.Background(), 30*time.Second)
 					defer ctxCancel()
@@ -907,7 +906,6 @@ func TestApplyClusterRoleBinding(t *testing.T) {
 	}
 
 	for _, tc := range tt {
-		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
@@ -920,7 +918,7 @@ func TestApplyClusterRoleBinding(t *testing.T) {
 			if tc.expectedErr != nil {
 				iterations = 1
 			}
-			for i := 0; i < iterations; i++ {
+			for i := range iterations {
 				t.Run("", func(t *testing.T) {
 					ctx, ctxCancel := context.WithTimeout(context.Background(), 30*time.Second)
 					defer ctxCancel()
@@ -1470,7 +1468,6 @@ func TestApplyRoleBinding(t *testing.T) {
 	}
 
 	for _, tc := range tt {
-		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
@@ -1483,7 +1480,7 @@ func TestApplyRoleBinding(t *testing.T) {
 			if tc.expectedErr != nil {
 				iterations = 1
 			}
-			for i := 0; i < iterations; i++ {
+			for i := range iterations {
 				t.Run("", func(t *testing.T) {
 					ctx, ctxCancel := context.WithTimeout(context.Background(), 30*time.Second)
 					defer ctxCancel()
