@@ -437,9 +437,7 @@ func TestValidateScyllaCluster(t *testing.T) {
 		},
 	}
 
-	for i := range tests {
-		test := tests[i]
-
+	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
 			t.Parallel()
 
