@@ -260,8 +260,7 @@ func TestIsStatefulSetRolledOut(t *testing.T) {
 		},
 	}
 
-	for i := range tt {
-		tc := tt[i]
+	for _, tc := range tt {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 			got, gotErr := IsStatefulSetRolledOut(tc.sts)
