@@ -171,7 +171,6 @@ key:
 	}
 
 	for _, tc := range tt {
-		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
@@ -310,8 +309,7 @@ func TestScyllaArguments(t *testing.T) {
 		},
 	}
 
-	for i := range ts {
-		test := ts[i]
+	for _, test := range ts {
 		t.Run(test.Name, func(t *testing.T) {
 			t.Parallel()
 
