@@ -42,9 +42,7 @@ func TestMax(t *testing.T) {
 			expectedMax: 9,
 		},
 	}
-	for i := range tt {
-		tc := tt[i]
-
+	for _, tc := range tt {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
